+++ packages/headless/src/ssr-next/commerce/factories/build-factory.ts
@@ -149,7 +149,7 @@ export const buildFactory =
     };
 
     if (options.onAccessTokenUpdate) {
-      options.onAccessTokenUpdate(updateEngineConfiguration);
+      options.onAccessTokenUpdate(updateEngineConfiguration, engine);
     }
 
     const controllers = buildControllerDefinitions({