+++ packages/headless/src/ssr/commerce/engine/commerce-engine.ssr.ts
@@ -57,8 +57,8 @@ export function defineCommerceEngine<
 
   const tokenManager = createAccessTokenManager(engineOptions.configuration.accessToken);
 
-  const onAccessTokenUpdate = (updateCallback: (accessToken: string) => void) => {
-    tokenManager.registerCallback(updateCallback);
+  const onAccessTokenUpdate = (updateCallback: (accessToken: string) => void, owner: object) => {
+    tokenManager.registerCallback(updateCallback, owner);
   };
 
   /**