+++ packages/quantic/force-app/main/default/lwc/quanticLoadMoreResults/e2e/quanticLoadMoreResults.e2e.ts
@@ -37,7 +37,6 @@ useCaseTestCases.forEach((useCase) => {
         });
       });
     });
-
     test.describe('when no results are available', () => {
       test.use({emptyResults: true});